+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 - update ruby to 3.4.2
 - update rails to 8.0.1
+- default console `update` for `format: :json` now returns a JSON with changes instead of an empty hash
 
 ## [6.2.5] - 2025-03-26